backend/requirements.txt
Django==4.2.7
gunicorn==21.2.0
gevent==23.9.1
boto3==1.34.0
botocore==1.34.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
